Fixed error in "drop fields" logic.
The fields to be dropped should persist through all of the other processing
performed on a given row.  Only *after* all the processing is done should the
marked fields be dropped from the record.  This allows the marked fields to be
used for other purposes, such as collecting lat/long information into a new
field, etc.

This also means that the dropped fields can be stored in a list rather than a
set or table or other data structure - after a row is processed, simply step
through the list and drop every field in the record with a matching name.