Added some custom date/time handling code.
The standard Date.toString() method puts in way too much information.  This is a
barer-bones version that shows just the date and the time, with no other extra
info.  The decision to change the Date prototype to use the new function was
mainly an experiment to see if it would work, but since this is the only point
in the code that uses date handling it's probably ok.