Added "zoom" and "unzoom" buttons for time slider.
The zoom button will re-establish the bounds of the slider to be the current
value range, and it will also store the old bounds in a stack.

The unzoom button pops a bounding range from the stack and applies it as the new
bounds, without changing the value of the sliders.  If there are no bounds in
the stack, the slider is unzoomed as far as it can go, and the unzoom button
will be disabled.