Functional simulation of data retrieval behavior.
When the user clicks the "update" button, the system currently generates some
random data, then passes it to the overlay so it can be drawn.

To make this fully functional, two things need to happen:

1. Instead of generating random test data, the retrieveData() function needs to
construct a MongoDB query out of the current state of the time slider, the
hashtag list, and possibly some kind of brushing primitives the user has drawn
on the map, issue that query to the database, and then pass the results, when
they arrive, to the overlay object.

2. The drawing routine must convert the lat/long data into mapspace coordinates.
(The test data is always generated in the [0,1]^2 range, so it is easy to draw
the points without using any lat/long transform logic)