Lat/long functionality mostly works.
This commit has the retrieveData() function statically supply the locations of
several well known cities by their lat/long coordinates.  This data flows to the
google maps overlay properly, and the glyphs are displayed in the proper place.
However, there is still a minor problem where the SVG element (visualized in
this commit by placing a translucent rect element within the SVG element when it
is created) is only the size of the map itself; when the user drags the map, the
SVG element is translated appropriately, and no markers outside of that original
region are visible.  This can probably be solved by always making the SVG
element extend across an entire copy of the whole-earth map, by computing the
pixel coordinates at which the upper left and lower right of the map lie, and
setting the SVG element to take up the whole space.

The next step is still to replace the example city data with an actual database
lookup.