Refactoring SSCI demo
Now the functional separation of components are a bit
more sane. There is a visCrossCat functional object that
contains the needed data manipulation and visualization
code. Also adding the raw CrossCat data files that are
loaded dynamically instead of relying on a Python script
to put all the data in one mega-object.