Fixing up things for Windows
Windows does not support symbolic links, so I moved
bootstrap-readable.css up a directory and removed
the bootstrap.css symbolic link.

Windows does not have "which" so instead we look
for the cherryd executable in CMake. We now configure
the tangelo control script to hard-code the python
shebang path and cherryd path.