The most basic possible scheme for implmenting .htaccess style http authentication.
This simply checks the containing directory of each request for a .htaccess
file.  If one is found, it "hot swaps" in a basic_auth instance to the
containing path.  This has the effect of immediately requiring the
username/password specified statically in tangelo.py itself (alice/scatter).

To make this really work, the server needs to keep a dict mapping paths to the
info found in that path's .htaccess file.  We need to think about things like,
do we want to track the mtime of the .htaccess files for the purpose of keeping
the server's dict up-to-date, etc.  A first attempt won't worry about that, but
will simply allow people to keep their applications relatively secure.