Correct pid file behavior when started as root and privileges are dropped.
Also added an option to NOT drop privileges when starting as root.

The os.remove() call has been put back in, because of the difficulties of
dealing with the pid file when dropping privileges.  For some reason, the
program creates the PID file as being owned by root even when privileges have
already been dropped.  The solution is to create the file preemptively, and then
drop its ownership manually.  For some reason this prevents the PIDFile plugin
from actually removing the file later, so we simply do it manually.  A bit of a
hack but it works cleanly.