Added Javascript unit testing infrastructure.
Jasmine is used to specify the unit tests, and PhantomJS runs the tests.  CMake
generates test suites by embedding various .js files found in
testing/js-unit-tests, into an HTML scaffold that includes code to take the test
results and report them nicely in HTML.  These files are placed in Tangelo's
default web space, and CTest generates invocations of PhantomJS to "run" the
tests using a modified run-jasmine.js.

Eventually, each test will need to depend on a test that launches the Tangelo
server.  There will be a similar teardown "teardown" test that will depend on
all of the other tests.  This will work somewhat differently on Windows, since
there isn't an option to launch Tangelo as a daemon there.