Closing standard streams in daemon child
This is proper daemonization procedure.  When we don't do this, it causes other
processes to believe that the tangelo control script may still be sending
output, and those processes can hang waiting on output that isn't coming.

This solves a problem with CTest running the tangelo script to control startup
and shutdown of a testing server instance.