Changing copy_if_different to copy
This prevents a timestamp issue: if a file is edited in such a way that it comes
to be the same as it was, its timestamp will nonetheless indicate it's a newer
file than its copy in the build tree.  On doing a "make", this causes the
copy_if_different action to fire, which will not copy the file (because they are
the same).  You get stuck in a slightly ugly state where make will not
"converge" to the null action, even though it has nothing to do.  Instead,
force-copying the file will keep the timestamps in sync and eliminate the
behavior.