Added mtime update to plugin refresh code
The plugins are now only refreshed if the plugin config file has been updated
since the last refresh operation.  This prevents, e.g., repeated attempts to
load failing plugins when the admin has not done anything to fix the issue in
the meantime.

This also fixes a slight problem with error reporting, and introduces the need
to solve #422.

Closes #419