Simplifying color log output
It is difficult to search for strings in the log output lines consisting of the
section (e.g., "ERROR") followed by the initial part of a message (e.g., "File
'blah' does not exist) as a single string (e.g., "ERROR File 'blah' does not
exist"), because the color escape sequences used to bracket each component
separately.  This removes the two sequences "in the middle", thus enabling this
type of string search, which is useful in writing tests, etc.