Adding a flake8 error to the ignore list
E402 complains about all module-level import statements not being at the
top of the file.  This is probably a good thing in general but seems to be
a new error added in a recent version of flake8 (as of right now, it is
not listed in the pep8 readthedocs documentation [0], for instance).

The error is arising now from ``startrek.py``, an example Python script
included in the SQLAlchemy tutorial in the docs.  Because simply fixing
the error in that file will cause it to go out of sync with the actual
tutorial, I am simply masking the error.

This commit blacklists the error for now, to replicate earlier build
behavior.  I will open an issue to remove it and fix the error in
``startrek.py`` and update the documentation to go along with it.

[0] http://pep8.readthedocs.org/en/latest/intro.html#error-codes