Import from source directory if in service area.
Add tests for importing locally and from tangelo.paths().

The paths computed in tangelo.paths are now more thread-safe and won't accrete if the same path is added multiple times.  There was a bug that if Tangelo were serving path foo/bar, then you could add a path foo/barbaz which wasn't in the served path.